hackernews/pyproject.toml
[tool.poetry]
name = "hackernews"
version = "0.1.0"
description = ""
authors = ["Christophe Blefari <christophe.blefari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.8"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
pandas = "^2.2.3"
google-cloud-bigquery = "^3.29.0"
google-auth = "^2.38.0"
pandas-gbq = "^0.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
